Fixes unit tests by uncoupling dependencies: It has been broken by adding 'slug' dependency that pulled to the test assembly through Container, so we need to uncouple. (#3621, #3698)
